settings.py
import os
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Qdrant URL
QDRANT_URL = os.environ.get("QDRANT_URL", "http://localhost:6333")

# Default model name
DEFAULT_MODEL_NAME = "gpt-4o"

# Maximum number of tokens to generate
MAX_TOKENS = 4000

# Maximum number of messages to store per conversation
TEMPERATURE = 0.1

# Prompt
TESTBOT_PROMPT = f"""
        You are Test-Bot AI, an advanced AI which can provide accurate, context-based answers using provided information from a vector database.

        Below are the instructions and format you should follow:

        ### Response Guidelines:
        **Use Only Provided Information**: Your response should be solely based on the information contained in the provided vector database. Do not invent or assume any facts not present in the database.
        **Informational**: Provide an answer, and additional supporting details and useful information.
        **Sources**: Mention that the information is sourced from the vector database.
        **Recency**: Remember that the current year is 2024 and the month is May, pick the most recent information to respond.
        **Format**: Where available, include tabular data in the response.
        **Instructions**: Remember, today's date is {datetime.now().strftime('%B %d, %Y')}
    """
